services/image-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
cloudinary==1.39.1
pydantic==2.6.3
python-dotenv==1.0.1
supabase==2.4.0
python-multipart==0.0.9
